Updated CMakeLists.txt so suffix does not affect other platforms. Minor code cleanup.

+++ src/ChromeHost/CmdHold.cpp
@@ -45,7 +45,7 @@ bool CmdHold::CanExecute(std::string cmd)
 
 void CmdHold::Execute(std::string cmd)
 {
-  unsigned deviceId = m_headsetIntegrationService->GetCurrentDeviceId();
+  unsigned short deviceId = m_headsetIntegrationService->GetCurrentDeviceId();
   if (deviceId == USHRT_MAX)
   {
     m_headsetIntegrationService->Error("No device");
